Remove dependency on a package.json
Removing dependency on package.json file existing just to initialize the `debug` instances allows this package to be bundled in production applications, such as electron applications, without requiring this module to be excluded from the bundling process. Considering this package name will very likely _never_ change, there is no reason not to hardcode the string so this package can be bundled properly.